requirements.txt
appnope==0.1.2
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
boto3==1.17.1
botocore==1.20.1
decorator==4.4.2
Django==3.1.6
django-address==0.2.5
django-ninja==0.10.2
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.14
psycopg2==2.8.6
ptyprocess==0.7.0
pycodestyle==2.6.0
pydantic==1.7.3
Pygments==2.7.4
pylint==2.6.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
wrapt==1.12.1
